pyproject.toml
[project]
name = "llm-council"
version = "0.1.1"
description = "Your LLM Council"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.3",
    "cryptography>=46.0.3",
    "requests>=2.32.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.12.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html",
    # Note: Coverage threshold set to 0 for now as we focus on critical functions
    # Target is 80% overall, but we're implementing incrementally
    # "--cov-fail-under=80",
]

[dependency-groups]
dev = [
    "pip-audit>=2.10.0",
    "pre-commit>=3.0.0",  # Optional: for git hooks (non-blocking)
    "ruff>=0.8.0",  # Formatter and linter for Python
    "bandit[toml]>=1.7.0",  # Security linter for Python
]

# Ruff configuration for formatting and linting
[tool.ruff]
# Line length matches .editorconfig
line-length = 100
target-version = "py310"

# Enable formatting
[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Skip magic trailing comma
skip-magic-trailing-comma = false
# Line ending style
line-ending = "auto"

# Linting rules
[tool.ruff.lint]
# Enable recommended rule sets
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (import sorting)
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]

# Ignore specific rules
ignore = [
    "E501",  # Line too long (handled by formatter)
    "B008",  # Do not perform function calls in argument defaults
]

# Exclude directories
exclude = [
    ".git",
    "__pycache__",
    "venv",
    ".venv",
    "dist",
    "build",
    "htmlcov",
    "node_modules",
]

# Import sorting configuration
[tool.ruff.lint.isort]
known-first-party = ["backend"]

# Bandit security linting configuration
[tool.bandit]
# Exclude directories and files
exclude_dirs = ["venv", ".venv", "__pycache__", "htmlcov", "dist", "build", "node_modules", "tests"]
# Skip specific test files
skips = ["B101"]  # Skip assert_used (assertions are fine in tests)
# Severity level: low, medium, high
severity_level = "medium"
# Confidence level: low, medium, high
confidence_level = "medium"
